requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.0=pyhd8ed1ab_0
aiohttp=3.10.5=py312h4389bb4_1
aiosignal=1.3.1=pyhd8ed1ab_0
annotated-types=0.7.0=pyhd8ed1ab_0
anyio=4.6.0=pyhd8ed1ab_1
asgiref=3.8.1=pypi_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=24.2.0=pyh71513ae_0
backoff=2.2.1=pypi_0
bcrypt=4.2.0=pypi_0
brotli-python=1.1.0=py312h275cf98_2
build=1.2.2=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.8.30=h56e8100_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py312h4389bb4_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
chroma=0.2.0=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.7=pypi_0
coloredlogs=15.0.1=pypi_0
deprecated=1.2.14=pypi_0
durationpy=0.7=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
fastapi=0.115.0=pypi_0
filelock=3.16.1=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=py312h4389bb4_1
fsspec=2024.9.0=pypi_0
google-auth=2.35.0=pypi_0
googleapis-common-protos=1.65.0=pypi_0
greenlet=3.1.1=py312h275cf98_0
grpcio=1.66.1=pypi_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.5=pyhd8ed1ab_0
httptools=0.6.1=pypi_0
httpx=0.27.2=pyhd8ed1ab_0
huggingface-hub=0.25.0=pypi_0
humanfriendly=10.0=pypi_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_0
importlib-metadata=8.4.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2024.2.1=h57928b3_1083
jiter=0.5.0=pypi_0
jsonpatch=1.33=pyhd8ed1ab_0
jsonpointer=3.0.0=py312h2e8e312_1
kubernetes=31.0.0=pypi_0
langchain=0.3.0=pyhd8ed1ab_0
langchain-chroma=0.1.4=pypi_0
langchain-community=0.3.0=pypi_0
langchain-core=0.3.5=pyhd8ed1ab_0
langchain-experimental=0.3.0=pypi_0
langchain-ollama=0.2.0=pypi_0
langchain-openai=0.2.0=pypi_0
langchain-text-splitters=0.3.0=pyhd8ed1ab_0
langgraph=0.2.23=pypi_0
langgraph-checkpoint=1.0.10=pypi_0
langsmith=0.1.125=pyhd8ed1ab_0
libblas=3.9.0=24_win64_mkl
libcblas=3.9.0=24_win64_mkl
libexpat=2.6.3=he0c23c2_0
libffi=3.4.4=hd77b12b_1
libhwloc=2.11.1=default_h8125262_1000
libiconv=1.17=hcfcfb64_2
liblapack=3.9.0=24_win64_mkl
libsqlite=3.46.1=h2466b09_0
libxcb=1.16=h013a479_1
libxml2=2.12.7=h0f24e4e_4
libzlib=1.3.1=h2466b09_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown-it-py=3.0.0=pypi_0
mdurl=0.1.2=pypi_0
mkl=2024.1.0=h66d3029_694
mmh3=5.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
msys2-conda-epoch=20160418=1
multidict=6.1.0=py312h4389bb4_0
mypy-extensions=1.0.0=pypi_0
numpy=1.26.4=py312h8753938_0
oauthlib=3.2.2=pypi_0
ollama=0.3.3=pypi_0
onnxruntime=1.19.2=pypi_0
openai=1.47.0=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.3.2=h2466b09_0
opentelemetry-api=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.27.0=pypi_0
opentelemetry-instrumentation=0.48b0=pypi_0
opentelemetry-instrumentation-asgi=0.48b0=pypi_0
opentelemetry-instrumentation-fastapi=0.48b0=pypi_0
opentelemetry-proto=1.27.0=pypi_0
opentelemetry-sdk=1.27.0=pypi_0
opentelemetry-semantic-conventions=0.48b0=pypi_0
opentelemetry-util-http=0.48b0=pypi_0
orjson=3.10.7=py312h426fad5_0
overrides=7.7.0=pypi_0
packaging=24.1=pyhd8ed1ab_0
pillow=10.4.0=pypi_0
pip=24.2=py312haa95532_0
posthog=3.6.6=pypi_0
protobuf=4.25.5=pypi_0
pthread-stubs=0.3=h3c9f919_1
pthreads-win32=2.9.1=hfa6e2cd_3
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pyclipper=1.3.0.post5=pypi_0
pycparser=2.22=pyhd8ed1ab_0
pydantic=2.9.2=pyhd8ed1ab_0
pydantic-core=2.18.4=pypi_0
pydantic-settings=2.5.2=pypi_0
pygments=2.18.0=pypi_0
pypdf=5.0.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
pyreadline3=3.5.4=pypi_0
pysocks=1.7.1=pyh0701188_6
python=3.12.6=hce54a09_1_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python_abi=3.12=5_cp312
pyyaml=6.0.2=py312h4389bb4_1
rapidocr-onnxruntime=1.3.24=pypi_0
regex=2024.9.11=pypi_0
requests=2.32.3=pyhd8ed1ab_0
requests-oauthlib=2.0.0=pypi_0
rich=13.8.1=pypi_0
rsa=4.9=pypi_0
setuptools=75.1.0=py312haa95532_0
shapely=2.0.6=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pyhd8ed1ab_0
sqlalchemy=2.0.35=py312h4389bb4_0
starlette=0.38.6=pypi_0
sympy=1.13.3=pypi_0
tbb=2021.13.0=hc790b64_0
tenacity=8.5.0=pyhd8ed1ab_0
tiktoken=0.7.0=pypi_0
tk=8.6.13=h5226925_1
tokenizers=0.20.0=pypi_0
typer=0.12.5=pypi_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
urllib3=2.2.3=pyhd8ed1ab_0
uvicorn=0.30.6=pypi_0
vc=14.40=h2eaa2aa_1
vc14_runtime=14.40.33810=ha82c5b3_21
vs2015_runtime=14.40.33810=h3bf8584_21
watchfiles=0.24.0=pypi_0
websocket-client=1.8.0=pypi_0
websockets=13.1=pypi_0
wheel=0.44.0=py312haa95532_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wrapt=1.16.0=pypi_0
xorg-kbproto=1.0.7=hcd874cb_1002
xorg-libx11=1.8.9=h0076a8d_1
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xorg-xextproto=7.3.0=hcd874cb_1003
xorg-xproto=7.0.31=hcd874cb_1007
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=h8ffe710_2
yarl=1.9.4=py312h4389bb4_1
zipp=3.20.2=pypi_0
zstandard=0.23.0=py312h7606c53_1
zstd=1.5.6=h0ea2cb4_0
